requirements.txt
asgiref==3.7.2
Babel==2.13.1
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap5==23.3
django-crispy-forms==2.1
django-money==3.4.0
money==1.3.0
py-moneyed==3.0
sqlparse==0.4.4
typing_extensions==4.8.0
